services/api_gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
pyjwt==2.8.0
cloudinary==1.36.0
python-multipart==0.0.6